invoice-service/requirements.txt
fastapi==0.100.0
uvicorn==0.22.0
sqlalchemy==2.0.17
pydantic==2.0.3
loguru==0.7.0
alembic==1.11.1
aio-pika==9.1.3
asyncpg==0.28.0
psycopg2-binary==2.9.6
httpx==0.24.1
python-multipart==0.0.6
PyJWT==2.7.0
python-dotenv==1.0.0
reportlab==3.6.12
